+++ src/0_utils.js
@@ -84,11 +84,11 @@ utils.cleanLinkData = function(linkData, config) {
 	if (WEB_BUILD) { // WTF, I don't know why I can't just do undef:false for the line.
 	/* jshint undef:true */
 		linkData['source'] = 'web-sdk';
-		if (linkData['data']['$desktop_url'] !== undefined) {
+		if (linkData['data'] && linkData['data']['$desktop_url'] !== undefined) {
 			linkData['data']['$desktop_url'] = linkData['data']['$desktop_url'].replace(/#r:[a-z0-9-_]+$/i, '').replace(/([\?\&]_branch_match_id=\d+)/, '');
 		}
 	}
-	linkData['data'] = goog.json.serialize(linkData['data']);
+	linkData['data'] = goog.json.serialize(linkData['data'] || {});
 	return linkData;
 };